feat: Integrate performance optimizations for large libraries
- Add optimized ApiService with request batching, caching, and deduplication
- Add DatabaseService with efficient pagination and caching for 300k+ items
- Replace LazyBookshelf with virtual scrolling implementation
- Remove temporary OPTIMIZED_* files from root directory
