--- OPTIMIZED_DATABASE.js
@@ -1,374 +1,0 @@
-/**
- * Optimized Database Operations for Large Libraries
- * Handles 300,000+ books efficiently with pagination and caching
- */
-
-class OptimizedAbsDatabase {
-  constructor() {
-    this.db = null
-    this.cache = new Map()
-    this.maxCacheSize = 1000 // Cache 1000 recent queries
-    this.batchSize = 500 // Process in batches
-  }
-
-  async initialize() {
-    // Initialize SQLite database with optimizations
-    this.db = await this.openDatabase()
-    await this.createIndexes()
-    await this.optimizeSettings()
-  }
-
-  async createIndexes() {
-    // Critical indexes for performance
-    const indexes = [
-      'CREATE INDEX IF NOT EXISTS idx_library_items_library_id ON library_items(libraryId)',
-      'CREATE INDEX IF NOT EXISTS idx_library_items_added_at ON library_items(addedAt DESC)',
-      'CREATE INDEX IF NOT EXISTS idx_library_items_title ON library_items(title)',
-      'CREATE INDEX IF NOT EXISTS idx_library_items_author ON library_items(author)',
-      'CREATE INDEX IF NOT EXISTS idx_library_items_series ON library_items(series)',
-      'CREATE INDEX IF NOT EXISTS idx_progress_library_item ON local_media_progress(libraryItemId)',
-      'CREATE INDEX IF NOT EXISTS idx_progress_updated ON local_media_progress(lastUpdate DESC)',
-      'CREATE INDEX IF NOT EXISTS idx_downloads_status ON downloads(status)',
-      
-      // Composite indexes for common queries
-      'CREATE INDEX IF NOT EXISTS idx_library_items_composite ON library_items(libraryId, addedAt DESC)',
-      'CREATE INDEX IF NOT EXISTS idx_library_items_filter ON library_items(libraryId, mediaType, isFinished)'
-    ]
-
-    for (const indexSql of indexes) {
-      try {
-        await this.executeQuery(indexSql)
-      } catch (error) {
-        console.warn('Failed to create index:', error)
-      }
-    }
-  }
-
-  async optimizeSettings() {
-    // SQLite performance optimizations
-    const optimizations = [
-      'PRAGMA synchronous = NORMAL',  // Faster than FULL, safer than OFF
-      'PRAGMA cache_size = -32000',   // 32MB cache
-      'PRAGMA temp_store = MEMORY',   // Use memory for temp storage
-      'PRAGMA mmap_size = 268435456', // 256MB memory map
-      'PRAGMA journal_mode = WAL',    // Write-Ahead Logging for better concurrency
-      'PRAGMA wal_autocheckpoint = 1000' // Checkpoint every 1000 pages
-    ]
-
-    for (const pragma of optimizations) {
-      try {
-        await this.executeQuery(pragma)
-      } catch (error) {
-        console.warn('Failed to apply optimization:', error)
-      }
-    }
-  }
-
-  /**
-   * Get library items with efficient pagination
-   */
-  async getLibraryItems(options = {}) {
-    const {
-      libraryId,
-      offset = 0,
-      limit = 100,
-      orderBy = 'addedAt',
-      orderDirection = 'DESC',
-      filter = {},
-      searchQuery = null
-    } = options
-
-    // Generate cache key
-    const cacheKey = JSON.stringify({ 
-      type: 'libraryItems', 
-      libraryId, 
-      offset, 
-      limit, 
-      orderBy, 
-      orderDirection, 
-      filter, 
-      searchQuery 
-    })
-
-    // Check cache first
-    if (this.cache.has(cacheKey)) {
-      return this.cache.get(cacheKey)
-    }
-
-    let sql = `
-      SELECT id, libraryId, data, lastUpdate
-      FROM library_items 
-      WHERE libraryId = ?
-    `
-    const params = [libraryId]
-
-    // Add filters
-    if (filter.mediaType) {
-      sql += ' AND JSON_EXTRACT(data, "$.mediaType") = ?'
-      params.push(filter.mediaType)
-    }
-
-    if (filter.isFinished !== undefined) {
-      sql += ' AND JSON_EXTRACT(data, "$.userMediaProgress.isFinished") = ?'
-      params.push(filter.isFinished ? 1 : 0)
-    }
-
-    if (filter.series) {
-      sql += ' AND JSON_EXTRACT(data, "$.media.metadata.series") LIKE ?'
-      params.push(`%${filter.series}%`)
-    }
-
-    // Add search
-    if (searchQuery) {
-      sql += ` AND (
-        JSON_EXTRACT(data, "$.media.metadata.title") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.authors") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.series") LIKE ?
-      )`
-      const searchPattern = `%${searchQuery}%`
-      params.push(searchPattern, searchPattern, searchPattern)
-    }
-
-    // Add ordering
-    const validOrderFields = ['addedAt', 'title', 'author', 'duration']
-    if (validOrderFields.includes(orderBy)) {
-      if (orderBy === 'title') {
-        sql += ` ORDER BY JSON_EXTRACT(data, "$.media.metadata.title") ${orderDirection}`
-      } else if (orderBy === 'author') {
-        sql += ` ORDER BY JSON_EXTRACT(data, "$.media.metadata.authors") ${orderDirection}`
-      } else if (orderBy === 'duration') {
-        sql += ` ORDER BY JSON_EXTRACT(data, "$.media.duration") ${orderDirection}`
-      } else {
-        sql += ` ORDER BY ${orderBy} ${orderDirection}`
-      }
-    }
-
-    // Add pagination
-    sql += ' LIMIT ? OFFSET ?'
-    params.push(limit, offset)
-
-    try {
-      const result = await this.executeQuery(sql, params)
-      
-      // Parse JSON data
-      const items = result.map(row => ({
-        ...row,
-        data: JSON.parse(row.data)
-      }))
-
-      // Cache result with TTL
-      this.cacheResult(cacheKey, items)
-      
-      return items
-    } catch (error) {
-      console.error('Failed to get library items:', error)
-      throw error
-    }
-  }
-
-  /**
-   * Get total count efficiently
-   */
-  async getLibraryItemCount(libraryId, filter = {}) {
-    const cacheKey = `count-${libraryId}-${JSON.stringify(filter)}`
-    
-    if (this.cache.has(cacheKey)) {
-      return this.cache.get(cacheKey)
-    }
-
-    let sql = 'SELECT COUNT(*) as count FROM library_items WHERE libraryId = ?'
-    const params = [libraryId]
-
-    // Add filters (same as above)
-    if (filter.mediaType) {
-      sql += ' AND JSON_EXTRACT(data, "$.mediaType") = ?'
-      params.push(filter.mediaType)
-    }
-
-    if (filter.isFinished !== undefined) {
-      sql += ' AND JSON_EXTRACT(data, "$.userMediaProgress.isFinished") = ?'
-      params.push(filter.isFinished ? 1 : 0)
-    }
-
-    try {
-      const result = await this.executeQuery(sql, params)
-      const count = result[0]?.count || 0
-      
-      // Cache count for 30 seconds
-      this.cacheResult(cacheKey, count, 30000)
-      
-      return count
-    } catch (error) {
-      console.error('Failed to get library count:', error)
-      return 0
-    }
-  }
-
-  /**
-   * Batch update progress for multiple items
-   */
-  async batchUpdateProgress(progressUpdates) {
-    if (!progressUpdates.length) return
-
-    const batchSize = this.batchSize
-    
-    try {
-      await this.executeQuery('BEGIN TRANSACTION')
-      
-      for (let i = 0; i < progressUpdates.length; i += batchSize) {
-        const batch = progressUpdates.slice(i, i + batchSize)
-        
-        const sql = `
-          INSERT OR REPLACE INTO local_media_progress 
-          (id, libraryItemId, currentTime, duration, progress, isFinished, lastUpdate)
-          VALUES ${batch.map(() => '(?,?,?,?,?,?,?)').join(',')}
-        `
-        
-        const params = batch.flatMap(update => [
-          update.id,
-          update.libraryItemId,
-          update.currentTime,
-          update.duration,
-          update.progress,
-          update.isFinished ? 1 : 0,
-          Date.now()
-        ])
-        
-        await this.executeQuery(sql, params)
-      }
-      
-      await this.executeQuery('COMMIT')
-      
-      // Invalidate related cache entries
-      this.invalidateCache('progress')
-      
-    } catch (error) {
-      await this.executeQuery('ROLLBACK')
-      console.error('Failed to batch update progress:', error)
-      throw error
-    }
-  }
-
-  /**
-   * Efficient search with full-text search
-   */
-  async searchLibraryItems(libraryId, query, options = {}) {
-    const { offset = 0, limit = 50 } = options
-    
-    const cacheKey = `search-${libraryId}-${query}-${offset}-${limit}`
-    
-    if (this.cache.has(cacheKey)) {
-      return this.cache.get(cacheKey)
-    }
-
-    // Use FTS if available, otherwise fall back to LIKE
-    const sql = `
-      SELECT id, libraryId, data, lastUpdate,
-             -- Ranking based on match type
-             CASE 
-               WHEN JSON_EXTRACT(data, "$.media.metadata.title") LIKE ? THEN 1
-               WHEN JSON_EXTRACT(data, "$.media.metadata.authors") LIKE ? THEN 2
-               WHEN JSON_EXTRACT(data, "$.media.metadata.series") LIKE ? THEN 3
-               ELSE 4
-             END as rank
-      FROM library_items 
-      WHERE libraryId = ? AND (
-        JSON_EXTRACT(data, "$.media.metadata.title") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.authors") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.series") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.narrators") LIKE ? OR
-        JSON_EXTRACT(data, "$.media.metadata.genres") LIKE ?
-      )
-      ORDER BY rank, JSON_EXTRACT(data, "$.media.metadata.title")
-      LIMIT ? OFFSET ?
-    `
-    
-    const searchPattern = `%${query}%`
-    const params = [
-      searchPattern, searchPattern, searchPattern, // For ranking
-      libraryId,
-      searchPattern, searchPattern, searchPattern, searchPattern, searchPattern, // For filtering
-      limit, offset
-    ]
-
-    try {
-      const result = await this.executeQuery(sql, params)
-      
-      const items = result.map(row => ({
-        ...row,
-        data: JSON.parse(row.data)
-      }))
-
-      this.cacheResult(cacheKey, items, 60000) // Cache for 1 minute
-      
-      return items
-    } catch (error) {
-      console.error('Failed to search library items:', error)
-      return []
-    }
-  }
-
-  /**
-   * Cache management
-   */
-  cacheResult(key, data, ttl = 300000) { // Default 5 minutes
-    // Implement LRU cache
-    if (this.cache.size >= this.maxCacheSize) {
-      const firstKey = this.cache.keys().next().value
-      this.cache.delete(firstKey)
-    }
-
-    this.cache.set(key, data)
-    
-    // Set TTL
-    if (ttl > 0) {
-      setTimeout(() => {
-        this.cache.delete(key)
-      }, ttl)
-    }
-  }
-
-  invalidateCache(pattern) {
-    for (const key of this.cache.keys()) {
-      if (key.includes(pattern)) {
-        this.cache.delete(key)
-      }
-    }
-  }
-
-  /**
-   * Database maintenance
-   */
-  async vacuum() {
-    try {
-      await this.executeQuery('VACUUM')
-      console.log('Database vacuum completed')
-    } catch (error) {
-      console.error('Failed to vacuum database:', error)
-    }
-  }
-
-  async analyze() {
-    try {
-      await this.executeQuery('ANALYZE')
-      console.log('Database analyze completed')
-    } catch (error) {
-      console.error('Failed to analyze database:', error)
-    }
-  }
-
-  // Base methods (implement according to existing Capacitor plugin)
-  async executeQuery(sql, params = []) {
-    // This would call the actual Capacitor plugin
-    // Implementation depends on existing database plugin
-    throw new Error('executeQuery must be implemented')
-  }
-
-  async openDatabase() {
-    // This would open the actual database
-    // Implementation depends on existing database plugin
-    throw new Error('openDatabase must be implemented')
-  }
-}
-
-export default OptimizedAbsDatabase
